+++ app/helpers/activity_helper.rb
@@ -10,4 +10,10 @@ module ActivityHelper
 
     presenter_position <= activity_position + 1
   end
+
+  def link_to_activity_parent(parent:, user:)
+    return if parent.nil?
+    return parent.title if parent.fund? && user.delivery_partner?
+    link_to parent.title, organisation_activity_path(parent.organisation, parent), {class: "govuk-link govuk-link--no-visited-state"}
+  end
 end